+++ src/index.js
@@ -3,11 +3,11 @@ import ReactDOM from "react-dom";
 import {createStore} from "redux";
 import {Provider} from "react-redux";
 import App from "./components/app/app.jsx";
-import {reducer} from "./reducer.js";
+import reducer from "./reducer/reducer";
 import {createAPI} from "./api";
 import {compose} from "redux";
 import {applyMiddleware} from "redux";
-import {Operation} from "./reducer";
+import {Operation} from "./reducer/data/data";
 import thunk from "redux-thunk";
 
 const api = createAPI(() => {});
