+++ src/reducer/data/selectors.js
@@ -1,0 +1,3 @@
+
+export const getCurrentMovie = (state) => state.data.currentMovie;
+export const getMovies = (state) => state.data.movies;
